# starlark: built-in types map of *Builtin