@afking 2020-03-05
Just for fun! I noticed the double allocation while trying to implement built ins. I’m happy to go with the simplified *Builtin.